--- server/egin/store.go
@@ -1,1 +1,0 @@
-package egin
